+++ app/src/main/java/com/example/mastertask/Models/UserModel.kt
@@ -79,7 +79,7 @@ class UserViewModel: ViewModel() {
                 user.habilidades = item.data!!["habilidades"] as List<Map<String?, Any?>>?
                 user.nome = item.data!!["nome"] as String?
                 user.numServicosFeitos = item.data!!["numServicosFeitos"] as Long?
-                user.somaAvaliacoes = item.data!!["somaAvaliacoes"] as Long?
+                user.somaAvaliacoes = item.data!!["somaAvaliacoes"] as Double?
                 users.add(user)
             }
 
@@ -102,7 +102,7 @@ class UserViewModel: ViewModel() {
             user.endereco = it.data!!["endereco"] as String?
             user.nome = it.data!!["nome"] as String?
             user.numServicosFeitos = it.data!!["numServicosFeitos"] as Long?
-            user.somaAvaliacoes = it.data!!["somaAvaliacoes"] as Long?
+            user.somaAvaliacoes = it.data!!["somaAvaliacoes"] as Double?
             getItemLiveData.setValue(user)
         }.addOnFailureListener {
             Log.d("get", it.localizedMessage!!)
